updating sinon
