+++ Gruntfile.js
@@ -304,6 +304,7 @@ module.exports = function (grunt) {
                 keepRunner: true,
                 vendor: [
                     'common/test/assets/javascripts/components/sinon/lib/sinon.js',
+                    'common/test/assets/javascripts/components/sinon/lib/sinon/call.js',
                     'common/test/assets/javascripts/components/sinon/lib/sinon/spy.js',
                     'common/test/assets/javascripts/components/sinon/lib/sinon/stub.js',
                     'common/test/assets/javascripts/components/sinon/lib/sinon/util/*.js',
