+++ common/test/assets/javascripts/bower.json
@@ -3,7 +3,7 @@
   "version": "0.1.0 ",
   "dependencies": {
     "jasmine": "1.3",
-    "sinon": "1.6",
+    "sinon": "1.7",
     "jasmine-sinon": "git@github.com:froots/jasmine-sinon.git",
     "seedrandom": "http://davidbau.com/encode/seedrandom.js"
   }
