+++ common/test/assets/javascripts/components/sinon/AUTHORS
@@ -5,6 +5,7 @@ Contributors
 Adrian Phinney, adrian.phinney@bellaliant.net
 August Lilleaas, august.lilleaas@gmail.com
 Ben Hockey, neonstalwart@gmail.com
+Brandon Heyer, brandonheyer@gmail.com
 Cormac Flynn, cormac.flynn@viadeoteam.com
 Cory, seeflanigan@gmail.com
 Domenic Denicola, domenic@domenicdenicola.com
@@ -18,12 +19,15 @@ Gord Tanner, gord@tinyhippos.com
 Gregers Gram Rygg, http://github.com/gregersrygg
 Irina Dumitrascu, me@dira.ro
 Jan Kopriva, jan.kopriva@gooddata.com
+Jonny Reeves, github@jonnyreeves.co.uk
 kbackowski, kbackowski@gmail.com
 Keith Cirkel, github@keithcirkel.co.uk
-Kevin Turner, kevin@decipherinc.com
-Konrad Holowinski, konrad.holowinski@gmail.com
 Kevin Decker, kpdecker@gmail.com
+Kevin Turner, kevin@decipherinc.com
+Kim Joar Bekkelund, kjbekkelund@gmail.com
+Konrad Holowinski, konrad.holowinski@gmail.com
 Luis Cardoso, luis.cardoso@feedzai.com
+Marten Lienen, marten.lienen@gmail.com
 Martin Brochhaus, mbrochh@gmail.com
 Martin Sander, forke@uni-bonn.de
 Maximilian Antoni, mail@maxantoni.de
@@ -32,6 +36,7 @@ Michael Jackson, mjijackson@gmail.com
 Olmo Maldonado, olmo.maldonado@gmail.com
 Raynos, raynos2@gmail.com
 Rodion Vynnychenko, roddiku@gmail.com
+Roman Potashow, justgook@gmail.com
 Scott Andrews, scothis@gmail.com
 Shawn Krisman telaviv@github
 Sven Fuchs, svenfuchs@artweb-design.de
@@ -42,6 +47,7 @@ Tim Ruffles, timruffles@googlemail.com
 Tobias Ebnöther, ebi@gorn.ch
 Tomás Corral Casas, amischol@gmail.com
 Tristan Koch, tristan.koch@1und1.de
+Victor Costan, costan@gmail.com
 Will Butler, will@butlerhq.com
 William Sears, MrBigDog2U@gmail.com
 Wesley Walser, waw325@gmail.com
