+++ common/test/assets/javascripts/components/sinon/build
@@ -48,7 +48,7 @@ buster-format not found, skipping. To build with buster-format support:
   contents = File.read(file)
 
   File.open(file, "w") do |f|
-    f.puts("var sinon = (function () {")
+    f.puts("this.sinon = (function () {")
     f.puts(File.read("./node_modules/buster-core/lib/buster-core.js"))
     f.puts(File.read("./node_modules/buster-format/lib/buster-format.js").gsub("var buster = this.buster || {};", ""))
     f.puts(contents)
@@ -87,8 +87,8 @@ Dir.chdir(File.dirname(__FILE__)) do
   merger << "lib/sinon/util/fake_server_with_clock.js"
   merger.save("pkg/sinon-server#{version_string}.js")
   add_license("pkg/sinon-server#{version_string}.js", version)
-  `cp #{output} pkg/sinon.js`
-  `cp pkg/sinon-ie#{version_string}.js pkg/sinon-ie.js`
+  FileUtils.cp(output, 'pkg/sinon.js')
+  FileUtils.cp("pkg/sinon-ie#{version_string}.js", 'pkg/sinon-ie.js')
 
   puts "Built Sinon.JS #{version}"
 end
