--- common/test/assets/javascripts/components/sinon/component.json
@@ -1,8 +1,0 @@
-{
-  "name": "sinon",
-  "version": "1.6.0",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/cjohansen/Sinon.JS.git"
-  }
-}
