+++ common/test/assets/javascripts/components/sinon/test/sinon.html
@@ -31,6 +31,7 @@
     <script src="../lib/sinon.js"></script>
     <script src="../lib/sinon/match.js"></script>
     <script src="../lib/sinon/spy.js"></script>
+    <script src="../lib/sinon/call.js"></script>
     <script src="../lib/sinon/stub.js"></script>
     <script src="../lib/sinon/mock.js"></script>
     <script src="../lib/sinon/assert.js"></script>
@@ -48,6 +49,7 @@
     <!-- Tests -->
     <script src="sinon_test.js"></script>
     <script src="sinon/spy_test.js"></script>
+    <script src="sinon/call_test.js"></script>
     <script src="sinon/stub_test.js"></script>
     <script src="sinon/mock_test.js"></script>
     <script src="sinon/match_test.js"></script>
